+++ tests/TestCase/Model/Table/UsersTableTest.php
@@ -222,6 +222,16 @@ class UsersTableTest extends TestCase
     }
 
     /**
+     * Test for `findAuth()` method
+     * @test
+     */
+    public function testFindAuth()
+    {
+        $query = $this->Users->find('auth');
+        $this->assertEquals('SELECT Users.id AS `Users__id`, Users.username AS `Users__username`, Users.password AS `Users__password`, Users.email AS `Users__email`, Users.active AS `Users__active`, Users.banned AS `Users__banned` FROM users Users', $query->sql());
+    }
+
+    /**
      * Test for `findBanned()` method
      * @test
      */